# make cuda call async, add cuda async timer