--- a/src/ngraph/runtime/gpu/cuda_emitter.cpp
+++ b/src/ngraph/runtime/gpu/cuda_emitter.cpp
@@ -206,6 +206,7 @@ size_t runtime::gpu::CUDAEmitter::build_pad(const runtime::gpu::GPURuntimeContex
                                           NULL, // shared mem and stream
                                           args_list,
                                           0)); // arguments
+                                          debug_sync();
         }});
     }
     else // pad value provided at compile time (static)
@@ -223,6 +224,7 @@ size_t runtime::gpu::CUDAEmitter::build_pad(const runtime::gpu::GPURuntimeContex
                                           NULL, // shared mem and stream
                                           args_list,
                                           0)); // arguments
+                                          debug_sync();
         }});
     }
 
@@ -323,6 +325,7 @@ size_t runtime::gpu::CUDAEmitter::build_pad_dynamic(const runtime::gpu::GPURunti
                                       NULL, // shared mem and stream
                                       args_list.data(),
                                       0)); // arguments
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(pad_dynamic));
@@ -411,6 +414,7 @@ size_t runtime::gpu::CUDAEmitter::build_slice(const runtime::gpu::GPURuntimeCont
                                       NULL, // shared mem and stream
                                       args_list.data(),
                                       0)); // arguments
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(kernel_runner));
@@ -493,6 +497,7 @@ size_t runtime::gpu::CUDAEmitter::build_reverse_sequence(const runtime::gpu::GPU
                                       NULL, // shared mem and stream
                                       args_list.data(),
                                       0)); // arguments
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(kernel_runner));
@@ -555,6 +560,7 @@ size_t runtime::gpu::CUDAEmitter::build_1d_max_pool(const GPURuntimeContext* ctx
                                           NULL, // shared mem and stream
                                           args_list,
                                           0)); // arguments
+                                          debug_sync();
         }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(pool));
@@ -743,6 +749,7 @@ size_t runtime::gpu::CUDAEmitter::build_avg_pool(const GPURuntimeContext* ctx,
                                           NULL,
                                           args_list,
                                           0));
+                                          debug_sync();
         }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(pool));
@@ -818,6 +825,7 @@ size_t runtime::gpu::CUDAEmitter::build_elementwise_n_to_1(const GPURuntimeConte
                                           NULL, // shared mem and stream
                                           args_list.data(),
                                           0)); // arguments
+                                          debug_sync();
         }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(ew));
@@ -953,6 +961,7 @@ size_t
                                       NULL,
                                       args_list.data(),
                                       0));
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(ew_collective));
@@ -1067,6 +1076,7 @@ size_t runtime::gpu::CUDAEmitter::build_reduce_window(const GPURuntimeContext* c
                                       NULL, // shared mem and stream
                                       args_list.data(),
                                       0)); // arguments
+                                      debug_sync();
 
     }});
 
@@ -1206,6 +1216,7 @@ size_t runtime::gpu::CUDAEmitter::build_replace_slice(const GPURuntimeContext* c
                                           NULL,
                                           args_list,
                                           0));
+                                          debug_sync();
         }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(replace_slice));
@@ -1317,6 +1328,7 @@ size_t runtime::gpu::CUDAEmitter::build_broadcast(const GPURuntimeContext* ctx,
                                       NULL,
                                       args_list,
                                       0));
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(broadcast));
@@ -1401,6 +1413,7 @@ size_t runtime::gpu::CUDAEmitter::build_reshape(const GPURuntimeContext* ctx,
                                       NULL,
                                       args_list,
                                       0));
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(reshape));
@@ -1653,6 +1666,7 @@ size_t runtime::gpu::CUDAEmitter::build_convolution(const GPURuntimeContext* ctx
                                       NULL,
                                       args_list,
                                       0));
+                                      debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(conv));
@@ -1806,4 +1820,18 @@ uint32_t runtime::gpu::CUDAEmitter::align_to_block_size(uint32_t threads, uint32
     }
     uint32_t r = (threads + block_size - 1) / block_size;
     return r;
+}
+
+void runtime::gpu::CUDAEmitter::sync()
+{
+    CUDA_SAFE_CALL(cuCtxSynchronize());
+    return;
+}
+
+void runtime::gpu::CUDAEmitter::debug_sync()
+{
+#ifdef NGRAPH_DEBUG_ENABLE
+    CUDA_SAFE_CALL(cuCtxSynchronize());
+#endif
+    return;
 }