--- a/src/ngraph/runtime/gpu/cudnn_emitter.cpp
+++ b/src/ngraph/runtime/gpu/cudnn_emitter.cpp
@@ -190,6 +190,7 @@ size_t runtime::gpu::CUDNNEmitter::build_reduce_forward(const runtime::gpu::GPUR
                                               beta,
                                               output_desc,
                                               outputs[0]));
+                                              debug_sync();
         }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(reduce));
@@ -242,6 +243,7 @@ size_t runtime::gpu::CUDNNEmitter::build_tensor_op(const GPURuntimeContext* ctx,
                                           beta_dt,
                                           descriptor,
                                           outputs[0]));
+                                          debug_sync();
         }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(tensor));
@@ -396,6 +398,7 @@ size_t runtime::gpu::CUDNNEmitter::build_convolution(const runtime::gpu::GPURunt
                                                 beta,
                                                 tensor_desc_1,
                                                 outputs[0]));
+                                                debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(conv));
@@ -473,6 +476,7 @@ size_t runtime::gpu::CUDNNEmitter::build_convolution_backward_data(
                                                      beta,
                                                      tensor_desc_1,
                                                      outputs[0]));
+                                                     debug_sync();
     }});
 
     primitive_index = this->m_primitive_emitter->insert(std::move(conv));
@@ -552,6 +556,7 @@ size_t runtime::gpu::CUDNNEmitter::build_convolution_backward_filter(
                                                        beta,
                                                        filter_desc,
                                                        outputs[0]));
+                                                       debug_sync();
     }});
     primitive_index = this->m_primitive_emitter->insert(std::move(conv));
     m_primitive_emitter->cache(hash, primitive_index);
@@ -646,6 +651,7 @@ size_t runtime::gpu::CUDNNEmitter::build_pooling(const runtime::gpu::GPURuntimeC
                                                     beta,
                                                     output_desc,
                                                     outputs[0]));
+                                                    debug_sync();
             }});
         break;
     }
@@ -676,6 +682,7 @@ size_t runtime::gpu::CUDNNEmitter::build_pooling(const runtime::gpu::GPURuntimeC
                                                      // adjoint of input
                                                      input_desc,
                                                      outputs[0]));
+                                                     debug_sync();
             }});
         break;
     }
@@ -742,6 +749,7 @@ size_t runtime::gpu::CUDNNEmitter::build_batchnorm(const runtime::gpu::GPURuntim
                                                                         inputs[3], // mean
                                                                         inputs[4], // variance
                                                                         epsilon));
+                                                                        debug_sync();
             }});
         break;
     }
@@ -779,6 +787,7 @@ size_t runtime::gpu::CUDNNEmitter::build_batchnorm(const runtime::gpu::GPURuntim
                                                                        epsilon,
                                                                        NULL,
                                                                        NULL));
+                                                                       debug_sync();
 
                 // convert to biased variance
                 CUDNN_SAFE_CALL(cudnnOpTensor(*ctx->cudnn_handle,
@@ -792,6 +801,7 @@ size_t runtime::gpu::CUDNNEmitter::build_batchnorm(const runtime::gpu::GPURuntim
                                               bias_factor,
                                               derived_param_desc,
                                               outputs[2]));
+                                              debug_sync();
             }});
         break;
     }
@@ -819,6 +829,7 @@ size_t runtime::gpu::CUDNNEmitter::build_batchnorm(const runtime::gpu::GPURuntim
                     epsilon,
                     NULL,   // inputs[3 /* mu batch mean*/],
                     NULL)); // inputs[4 /* 1/sig**2 batch inverse variance*/]);
+                    debug_sync();
             }});
         break;
     }
@@ -871,6 +882,7 @@ size_t runtime::gpu::CUDNNEmitter::build_softmax(const runtime::gpu::GPURuntimeC
                                                 beta,
                                                 tensor_desc,
                                                 outputs[0]));
+                                                debug_sync();
         }});
         break;
     }
@@ -888,6 +900,7 @@ size_t runtime::gpu::CUDNNEmitter::build_softmax(const runtime::gpu::GPURuntimeC
                                                  beta,
                                                  tensor_desc,
                                                  outputs[0]));
+                                                 debug_sync();
         }});
         break;
     }
@@ -896,4 +909,18 @@ size_t runtime::gpu::CUDNNEmitter::build_softmax(const runtime::gpu::GPURuntimeC
     primitive_index = this->m_primitive_emitter->insert(std::move(softmax));
     m_primitive_emitter->cache(hash, primitive_index);
     return primitive_index;
+}
+
+void runtime::gpu::CUDNNEmitter::sync()
+{
+    CUDA_RT_SAFE_CALL(cudaDeviceSynchronize());
+    return;
+}
+
+void runtime::gpu::CUDNNEmitter::debug_sync()
+{
+#ifdef NGRAPH_DEBUG_ENABLE
+    CUDA_RT_SAFE_CALL(cudaDeviceSynchronize());
+#endif
+    return;
 }