--- a/src/ngraph/runtime/gpu/cudnn_emitter.hpp
+++ b/src/ngraph/runtime/gpu/cudnn_emitter.hpp
@@ -123,6 +123,9 @@ namespace ngraph
                                      const Prop& direction,
                                      const Shape& tensor_shape);
 
+                void debug_sync();
+                void sync();
+
             private:
                 CUDNNEmitter(GPUPrimitiveEmitter* emitter);