Found cuda version `negative` is 6x slower than cudnn version.  By dig into the problem I found cudnn is running on async mode but cuda is running on sync mode. So the cpu timer for `timing-detail` does not make sense. Here we can both cuda call and gpu timer to support async mode.

Why we change to async calls.
Performance 
some reference here:
https://www.cs.virginia.edu/~mwb7w/cuda_support/kernel_overhead.html

Why we can change to async calls.
1. since currently all ops running on GPU in default stream, it guaranteed to be run in order. 
2. the only time need to sync is the data copy between device and host. and we are currently using sync copy(I could be wrong here, need to verify).

Result (since only the total time is correct for previous version, we only compare that)
- before  without timing-detail:
```
enchmarking test/models/mxnet/LSTM_backward.json, GPU backend, 10 iterations.
compile time: 13,936ms
77.3ms per iteration

---- Aggregate times per op type ----

---- Aggregate times per op type/shape ----
```
- after without timing-detail:
```
Benchmarking test/models/mxnet/LSTM_backward.json, GPU backend, 10 iterations.
compile time: 13,603ms
61.4ms per iteration

---- Aggregate times per op type ----

---- Aggregate times per op type/shape/count ----
```

`timing-detail` after change, a `count` is been add to the list after shape, which shows the total number of the op run in each iteration. 

- Note that `timing-detail` will make both compile and run slower, but it will be useful if we compare apple to apple for future performance optimization:
```
Benchmarking test/models/mxnet/LSTM_backward.json, GPU backend, 10 iterations.
compile time: 28,752ms
87.9ms per iteration

---- Aggregate times per op type ----
ReplaceSlice     15,388us
Multiply         10,041us
Reshape           9,206us
Add               8,495us
Dot               8,077us
Sum               4,615us
Negative          2,910us
Divide            2,592us
Slice             2,496us
Broadcast         1,040us
Cosh                981us
Result               95us

---- Aggregate times per op type/shape/count ----
Multiply {32, 200} 1678           10,041us
ReplaceSlice {32, 60, 200} 60      9,848us
ReplaceSlice {32, 800} 480         5,540us
Add {32, 200} 596                  3,735us
Dot {200, 32} 238                  3,641us
Sum {800} 240                      3,480us
Negative {32, 200} 720             2,910us
Add {800, 200} 236                 2,796us
Divide {32, 200} 600               2,592us
Dot {200, 800} 240                 2,586us
Reshape {800, 200} 240             1,953us
Reshape {800, 32} 238              1,839us
Slice {32, 200} 360                1,811us
Reshape {32, 200} 298              1,782us
Add {800} 236                      1,606us
Reshape {200, 800} 240             1,491us
Reshape {200, 32} 240              1,409us
Sum {10000} 1                      1,135us
Cosh {32, 200} 240                   981us
Broadcast {32, 800} 120              953us
Dot {200, 10000} 1                   935us
Dot {200, 1920} 1                    915us
Slice {32, 1, 200} 119               685us
Reshape {10000, 1920} 1              366us
Add {32, 1, 200} 59                  358us
Reshape {32, 1, 200} 60              246us
Broadcast {32, 60, 200} 1             87us
Reshape {200, 10000} 1                43us
Result {10000, 200} 1                 42us
Reshape {10000, 200} 1                41us
Result {800, 200} 4                   24us
Result {800} 4                        16us
Reshape {1920, 200} 1                 13us
Reshape {200, 1920} 1                 12us
Reshape {32, 60, 200} 1               11us
Result {32, 60, 200} 1                 8us
Result {10000} 1                       5us
```