@fengleitian 2018-07-12
@csullivan good point, we can add sync in debug mode.